Moved everything from Primitives to Miscellaneous. Will begin adding styles from Typography to Primitives category
